feat: move import statements to the top
Import statements should not be in the middle of the code, but on top,
so if autoimport finds anyone, it will move it.

This is useful to write the import statement wherever you are, so you
don't have to go up to the beginning of the file each time autoimport is
not able to detect a missing import.

style(services): organize the functions.

So they can be read upside down.