fix: import developing package objects when not in src directory
When executed in the tests directory, autoimport was not able to import
the package objects because there was an error finding the project root
directory.

feat: import objects defined in the __all__ special variable

docs: warn that the object finding in developing package may fail.

Given that you execute `autoimport` inside a virtualenv where the
package is installed with `pip install -e .`, when there is an import
error in a file that is indexed in the package, `autoimport` won't be
able to read the package contents as the `import` statement will fail.
So it's a good idea to run autoimport from a virtualenv that has
a stable version of the package we are developing.

refactor: remove the unused path attribute of SourceCode